+++ modules/dxp/apps/portal-workflow/portal-workflow-metrics-web/src/main/resources/META-INF/resources/js/components/process-metrics/filter/store/AssigneeStore.es.js
@@ -15,8 +15,8 @@ import React, {createContext, useContext, useEffect, useState} from 'react';
 import {buildFallbackItems} from '../../../../shared/components/filter/util/filterEvents.es';
 import {ErrorContext} from '../../../../shared/components/request/Error.es';
 import {LoadingContext} from '../../../../shared/components/request/Loading.es';
+import {usePrevious} from '../../../../shared/hooks/usePrevious.es';
 import {compareArrays} from '../../../../shared/util/array.es';
-import {usePrevious} from '../../../../shared/util/hooks.es';
 import {AppContext} from '../../../AppContext.es';
 
 const useAssignee = (assigneeKeys, processId) => {
