+++ modules/dxp/apps/portal-workflow/portal-workflow-metrics-web/src/main/resources/META-INF/resources/js/components/process-metrics/filter/store/CustomTimeRangeStore.es.js
@@ -16,7 +16,7 @@ import {
 	parse,
 	stringify
 } from '../../../../shared/components/router/queryString.es';
-import {useRouter} from '../../../../shared/components/router/useRouter.es';
+import {useRouter} from '../../../../shared/hooks/useRouter.es';
 import moment from '../../../../shared/util/moment.es';
 import {
 	formatDateEnLocale,
