+++ modules/dxp/apps/portal-workflow/portal-workflow-metrics-web/src/main/resources/META-INF/resources/js/components/process-metrics/filter/store/SLAStatusStore.es.js
@@ -13,8 +13,8 @@
 import React, {createContext, useEffect, useState} from 'react';
 
 import {buildFallbackItems} from '../../../../shared/components/filter/util/filterEvents.es';
+import {usePrevious} from '../../../../shared/hooks/usePrevious.es';
 import {compareArrays} from '../../../../shared/util/array.es';
-import {usePrevious} from '../../../../shared/util/hooks.es';
 
 const slaStatusConstants = {
 	onTime: 'OnTime',
